ads-datakatalog-func/requirements.txt
azure-functions
azure-identity
azure-purview-catalog
datacatalogtordf
cachetools
types-cachetools